incluimos sweetalert2

+++ app/Http/Controllers/Admin/FamilyController.php
@@ -38,6 +38,12 @@ class FamilyController extends Controller
 
         Family::create($request->all());
 
+        session()->flash('swal', [
+            'icon' => 'success',
+            'title' => '¡Buen trabajo!',
+            'text' => 'Familia creada correctamente.',
+        ]);
+
         return redirect()->route('admin.families.index');
     }
 
@@ -68,6 +74,12 @@ class FamilyController extends Controller
 
         $family->update($request->all());
 
+        session()->flash('swal', [
+            'icon' => 'success',
+            'title' => '¡Buen trabajo!',
+            'text' => 'Familia actualizada correctamente.',
+        ]);
+
         return redirect()->route('admin.families.edit', $family);
     }
 
@@ -76,7 +88,24 @@ class FamilyController extends Controller
      */
     public function destroy(Family $family)
     {
+
+        if($family->categories->count() > 0) {
+            session()->flash('swal', [
+                'icon' => 'error',
+                'title' => '¡Ups!',
+                'text' => 'No puedes eliminar una familia que tiene categorías asociadas.',
+            ]);
+
+            return redirect()->route('admin.families.edit', $family);
+        }
+
         $family->delete();
+
+        session()->flash('swal', [
+            'icon' => 'success',
+            'title' => '¡Buen trabajo!',
+            'text' => 'Familia eliminada correctamente.',
+        ]);
 
         return redirect()->route('admin.families.index');
     }
